+++ src/types/rx-storage.d.ts
@@ -2,6 +2,7 @@ import type { ChangeEvent } from 'event-reduce-js';
 import { RxDocumentMeta } from './rx-document';
 import { MangoQuery } from './rx-query';
 import { RxJsonSchema } from './rx-schema';
+import { StringKeys } from './util';
 
 /**
  * The document data how it comes out of the storage instance.
@@ -333,3 +334,40 @@ export type RxStorageDefaultCheckpoint = {
     id: string;
     lwt: number;
 } | null;
+
+
+
+export type CategorizeBulkWriteRowsOutput<RxDocType> = {
+    bulkInsertDocs: BulkWriteRow<RxDocType>[];
+    bulkUpdateDocs: BulkWriteRow<RxDocType>[];
+    /**
+     * Ids of all documents that are changed
+     * and so their change must be written into the
+     * sequences table so that they can be fetched via
+     * RxStorageInstance().getChangedDocumentsSince().
+     */
+    changedDocumentIds: RxDocumentData<RxDocType>[StringKeys<RxDocType>][];
+
+    /**
+     * TODO directly return a docId->error object
+     * like in the return value of bulkWrite().
+     * This will improve performance because we do not have to iterate
+     * over the error array again.
+     */
+    errors: RxStorageBulkWriteError<RxDocType>[];
+    eventBulk: EventBulk<RxStorageChangeEvent<RxDocumentData<RxDocType>>, any>;
+    attachmentsAdd: {
+        documentId: string;
+        attachmentId: string;
+        attachmentData: RxAttachmentWriteData;
+    }[];
+    attachmentsRemove: {
+        documentId: string;
+        attachmentId: string;
+    }[];
+    attachmentsUpdate: {
+        documentId: string;
+        attachmentId: string;
+        attachmentData: RxAttachmentWriteData;
+    }[];
+};